committing change for into to long change for total size and file size

+++ src/main/java/fr/gouv/culture/francetransfert/application/resources/model/DirectoryRepresentation.java
@@ -11,7 +11,7 @@ import java.util.List;
 @NoArgsConstructor
 public class DirectoryRepresentation extends DataRepresentation {
 
-    private int totalSize;
+    private long totalSize;
 
     @Valid
     private List<FileRepresentation> files;
